src/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
# basic dependencies
lightning==2.4.0
lightning_utilities==0.11.8
torch==2.2.0
torchmetrics==1.6.0
torchvision==0.17.0
transformers==4.40.0
tokenizers==0.19.1
deepspeed==0.13.1
accelerate==0.26.1
peft==0.4.0
timm==1.0.3
numpy==1.24.4
pandas==2.2.2
scipy==1.14.1
scikit_learn==1.5.0
# data processing
decord==0.6.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
pillow==10.4.0
Pillow_SIMD==9.5.0.post2
scenedetect==0.6.3
opencv-python==4.6.0.66
pysubs2
# misc
omegaconf==2.3.0
scikit-learn==1.2.2
huggingface_hub==0.23.4
hydra_core==1.3.2
sentencepiece==0.1.99
shortuuid
einops==0.6.1
einops-exts==0.0.4
bitsandbytes==0.43.0
pydantic>=2.0
markdown2[all]
gradio==3.50.0
gradio_client==0.6.1
httpx==0.24.1
requests
openai==1.51.2
uvicorn==0.31.1
fastapi==0.115.0
tensorboard==2.18.0
tensorboard_data_server==0.7.2
tqdm==4.66.5
wandb==0.18.3
tabulate==0.9.0
spaces==0.29.2
